Initial AI setup
This adds a new namespace: AI. inside this there is a computerPlayer
class that holds information for the computer player and allows it to
make moves. It still needs some functions built out, and to be tested. I
am going to need some help connecting this with the program.

It also adds the Board class to the TICSET namespace. Usman, please
check the comments, as I will need some help integrating that with the
login.

-Joe